TIP-303: move Jobfactory from BatchBundle to component
