+++ src/Akeneo/Bundle/BatchBundle/Resources/config/services.yml
@@ -1,7 +1,7 @@
 parameters:
     akeneo_batch.connectors.class:                            Akeneo\Bundle\BatchBundle\Connector\ConnectorRegistry
     akeneo_batch.job_repository.class:                        Akeneo\Bundle\BatchBundle\Job\DoctrineJobRepository
-    akeneo_batch.job_factory.class:                           Akeneo\Bundle\BatchBundle\Job\JobFactory
+    akeneo_batch.job_factory.class:                           Akeneo\Component\Batch\Job\JobFactory
     akeneo_batch.step_factory.class:                          Akeneo\Bundle\BatchBundle\Step\StepFactory
     akeneo_batch.job_instance_factory.class:                  Akeneo\Bundle\BatchBundle\Job\JobInstanceFactory
     akeneo_batch.logger_subscriber.class:                     Akeneo\Bundle\BatchBundle\EventListener\LoggerSubscriber
